Adds ability to authenticate a password combo.
Authenticates passwords using UUID and SHA512 password hashing.
The class will store the salt, created by the UUID.
The username and password will be hashed that cannot be reversed so that
the username must be hashed each time it is authenticated.
The password and username cannot be unhashed.

Returns 1 if successful
Returns 0 if the name was already taken.