Add ablility to ID and change transaction
whoami is an accessor method that gets the userID_owner. Changed the way
customer handles transaction. Now, it is requested delta change rather
than immediate transaction. requestDelta is the method in which it
appends the delta value to be accepted. It will append the value and the
date at which it was requested into the list of __requests.